[Modify] login, signup function - require v4 수정 - TimeStamep export 추가 - SignUp::UpdateUser -> 덮어쓰기 함수 추가(writeUser()). 지금까지 메모리에 저장만 하고 json은 변경 안함 - loginfetch:: 이메일 유효성 검사 -> loginEmailCheck로 분류
