+++ community/JS/Login/TimeStamp.js
@@ -2,3 +2,5 @@ function getTime(){
     const now = new Date();
     return now;
 }
+
+module.exports = {getTime};
